+++ lib/voicemail/matcher.rb
@@ -1,3 +1,6 @@
+# encoding: utf-8
+# This template exists for the purpose of being overridden.  See the README for more information.
+
 module Voicemail
   class Matcher
     def initialize(entered, actual)
@@ -6,8 +9,6 @@ module Voicemail
     end
 
     def matches?
-      logger.info "input : #{@entered}"
-      logger.info "actual: #{@actual}"
       @entered == @actual
     end
   end