+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   * CHANGE - Menu intros are one message, making them skippable.
   * FEATURE - Personalized voicemail greetings can be deleted
   * FEATURE - Add message to play after recording a voicemail is complete
+  * FEATURE - Add an option to erase all messages from a mailbox
+  * FEATURE - Put a timeout on waiting for a PIN
+  * FEATURE - Allow messages to be skipped
   * BUG - Resolve issue where caller presses the pound key twice in succession
 
 #v1.0.0 - 2014-01-02