+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+#develop
+  * CHANGE - Menu intros are one message, making them skippable.
+  * FEATURE - Personalized voicemail greetings can be deleted
+  * FEATURE - Add message to play after recording a voicemail is complete
+  * FEATURE - Add an option to erase all messages from a mailbox
+  * FEATURE - Put a timeout on waiting for a PIN
+  * FEATURE - Allow messages to be skipped
+  * BUG - Resolve issue where caller presses the pound key twice in succession
+
 #v1.0.0 - 2014-01-02
   * FEATURE - Optional per-mailbox override of default recording hash
   * FEATURE - Optional rerecording of voicemail messages