+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-#develop
+# develop
+  * FEATURE - Refactor voicemail storage (backward incompatible change)
+
+# v1.1.0 - 2014-05-07
   * CHANGE - Menu intros are one message, making them skippable.
   * FEATURE - Personalized voicemail greetings can be deleted
   * FEATURE - Add message to play after recording a voicemail is complete
@@ -7,7 +10,7 @@
   * FEATURE - Allow messages to be skipped
   * BUG - Resolve issue where caller presses the pound key twice in succession
 
-#v1.0.0 - 2014-01-02
+# v1.0.0 - 2014-01-02
   * FEATURE - Optional per-mailbox override of default recording hash
   * FEATURE - Optional rerecording of voicemail messages
   * FEATURE - I18n numeric methods